pyproject.toml
[project]
name = "grafeo-llamaindex"
version = "0.1.1"
description = "LlamaIndex PropertyGraphStore backed by GrafeoDB embedded graph database"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.12"
authors = [{ name = "GrafeoDB" }]
keywords = ["llamaindex", "graph-rag", "property-graph", "grafeo", "knowledge-graph"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Database :: Database Engines/Servers",
]
dependencies = [
    "grafeo>=0.4.4",
    "llama-index-core>=0.14,<1",
]

[project.urls]
Homepage = "https://github.com/GrafeoDB/grafeo-llamaindex"
Source = "https://github.com/GrafeoDB/grafeo-llamaindex"
Issues = "https://github.com/GrafeoDB/grafeo-llamaindex/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/grafeo_llamaindex"]

# ── LlamaHub metadata (for discovery) ────────────────────────────────────────────────────────────
[tool.llamahub]
contains_example = true
import_path = "grafeo_llamaindex"

[tool.llamahub.class_authors]
GrafeoPropertyGraphStore = "GrafeoDB"

# ── Ruff ──────────────────────────────────────────────────────────────────────────────────────────
[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]

# ── Ty ────────────────────────────────────────────────────────────────────────────────────────────
[tool.ty.environment]
python-version = "3.12"

# ── Pytest ────────────────────────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=grafeo_llamaindex --cov-report=term-missing"

# ── Coverage ─────────────────────────────────────────────────────────────────────────────────────
[tool.coverage.run]
source = ["grafeo_llamaindex"]
omit = ["*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

# ── UV ────────────────────────────────────────────────────────────────────────────────────────────
[tool.uv]
dev-dependencies = [
    "pytest>=8",
    "pytest-cov>=7",
    "ruff>=0.11",
    "ty>=0.0.10",
    "llama-index>=0.12",
]
